Remove addrinfo caching mechanism
This is a part of the previous commit. Due to the property
of domain name system, I've decided that users must provide
IP address themselves, but still provide a shorthand for
ping-ing only one host.
My consideration is, this decreases the complexity of this
library, but also give users the freedom to decide what
exact address to ping.